refactor(web-server): requestHandler to file
Move the static file handler code to a separate file for code
readability.
